+++ client/src/BoothMap/functions/firestore.js
@@ -1,0 +1,3 @@
+import firebase from '../../firebase'
+var boothsRef = firebase.firestore().collection("booths");
+export default boothsRef
